Pluck out the attributes of the mount point and pass them as props
This allows us to generate a mount point at jekyll-build-time to pass in
the frontmatter as data-attributes. Then, in the react app, when we do
the mounting of the component, we use basic JS to pluck out the
attributes on the mount element and pass them as props to the react
element.

This allows an earlier bake-in-vars-from-frontmatter to get to the react
component at run-later-on-browser time